flag handling and updated final report

+++ cleaner/actions.py
@@ -1,13 +1,13 @@
 # handle what we actually do to the files -- delete, archive
-import sys
 import os
 # given two files, one of which is the original and the other a clone, delete the clone and add a softlink to the original
 # note that the parameters are FileMetadata objects
-def delete_file(original_file, duplicate_file):
+def delete_file(original_file, duplicate_file, dry_run):
     failed_links = 0
     try:
-        os.remove(duplicate_file.path)
-        os.symlink(src=original_file.path, dst=duplicate_file.path)
+        if not dry_run:
+            os.remove(duplicate_file.path)
+            os.symlink(src=original_file.path, dst=duplicate_file.path)
         print(f"Deleting {duplicate_file.path} and linking to {original_file.path}")
     except Exception as e:
         print(f"Failed to create symlink from {duplicate_file.path} to {original_file.path}: {e}")
@@ -17,7 +17,7 @@ def delete_file(original_file, duplicate_file):
 
 # given a list of FileMetadata Objects, all of which are duplicates of each other, find the one with the most recent timestamp/mtime
 # delete the rest, and return a list containing the deleted files for records.py
-def delete_dupes(file_lst):
+def delete_dupes(file_lst, dry_run):
     failed_links = 0
     # Find the file with the most recent modification time
     most_recent_file = max(file_lst, key=lambda f: f.timestamp)
@@ -25,5 +25,5 @@ def delete_dupes(file_lst):
     # pop it, and delete everything else
     file_lst.remove(most_recent_file)
     for file in file_lst:
-        failed_links += delete_file(most_recent_file, file)
+        failed_links += delete_file(most_recent_file, file, dry_run)
     return file_lst, failed_links # return all the files that have been deleted, for bookeeping
